Fold the Gemma shard-mode tests into the shared decoder-layer test file
gemma_layers_test.py duplicated the config table, the mesh helper and the
parameterization already in decoder_layer_model_mode_test.py for a second
per-layer invariant. Move the two shard-mode tests in as a second class there
and drop the standalone file; the module now reads as "invariants every decoder
layer has to satisfy", with the model-mode and shard-mode checks side by side.

The shard-mode class builds its mesh with get_mesh_from_config rather than the
file's _mesh helper, which types the axes Auto and so is rejected by reshard.

Also pin TestNnxConstructNamedSharding to a size-one mesh. It asserts on which
axis *name* a logical axis resolves to and uses odd-shaped arrays, so sizing the
mesh off the ambient device count made it fail as soon as another cpu_only
module in the same pytest process raised the CPU device count -- which
decoder_layer_model_mode_test.py does, to give the shard-mode tests an fsdp axis
that actually splits.
