+++ tests/unit/decoder_layer_model_mode_test.py
@@ -12,34 +12,55 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""A decoder layer must attend in the mode it is called with, not the one it was built in.
+"""Invariants every decoder layer has to satisfy, checked one layer at a time.
 
-Under NNX a model object is built once and reused across modes — MaxEngine builds it
-in PREFILL and then passes MODEL_MODE_AUTOREGRESSIVE per call while decoding. A layer
-that forwards its construction-time `self.model_mode` to attention instead of the
-`model_mode` argument therefore runs the prefill KV-cache path during decode and
-produces wrong tokens from the first generated one, with no error anywhere.
+`DecoderLayerModelModeTest` — a layer must attend in the mode it is called with, not
+the one it was built in. Under NNX a model object is built once and reused across
+modes: MaxEngine builds it in PREFILL and then passes MODEL_MODE_AUTOREGRESSIVE per
+call while decoding. A layer that forwards its construction-time `self.model_mode` to
+attention instead of the `model_mode` argument therefore runs the prefill KV-cache path
+during decode and produces wrong tokens from the first generated one, with no error
+anywhere. The check stops at the attention boundary and asserts on the mode that
+arrives there, so it stays cheap and does not depend on any particular attention
+implementation.
 
-The check stops at the attention boundary and asserts on the mode that arrives there,
-so it stays cheap and does not depend on any particular attention implementation.
+`DecoderLayerShardModeTest` — a layer must produce the same activations under both
+shard modes. Under `shard_mode=auto` a missing sharding annotation costs nothing:
+GSPMD infers a layout. Under `shard_mode=explicit` the same omission is a hard
+`ShardingTypeError`, because JAX type-checks the sharding of every operation, so each
+layer has to hand its sub-modules an `out_sharding` (and, for the MLP, an
+`intermediate_sharding`) rather than rely on inference. A regression shows up either as
+an exception from the explicit run or as a numerical difference, which is what a wrong
+— as opposed to merely absent — annotation produces. Only the layers already on the
+explicit-sharding allowlist in `configs/types.py` are listed.
 """
 
+import os
 import sys
 import unittest
 from unittest import mock
 
+# A single-device mesh shards nothing, so the shard-mode tests below would pass even
+# with every annotation removed. Ask XLA for eight host devices before JAX initializes
+# its backend so the fsdp axis actually splits the activations.
+os.environ.setdefault("XLA_FLAGS", "--xla_force_host_platform_device_count=8")
+
+# pylint: disable=wrong-import-position
 import jax
 import jax.numpy as jnp
+import numpy as np
 from absl.testing import parameterized
 from flax import nnx
 from flax.linen import partitioning as nn_partitioning
 
-from maxtext.common.common_types import MODEL_MODE_AUTOREGRESSIVE, MODEL_MODE_PREFILL
+from maxtext.common.common_types import MODEL_MODE_AUTOREGRESSIVE, MODEL_MODE_PREFILL, MODEL_MODE_TRAIN, ShardMode
 from maxtext.configs import pyconfig
 from maxtext.layers import attention_mla, attentions
-from maxtext.models import deepseek, gemma, gemma2, llama2, mistral, qwen2, qwen3
+from maxtext.models import deepseek, gemma, gemma2, gemma3, llama2, mistral, qwen2, qwen3
 from maxtext.utils import maxtext_utils
 from tests.utils.test_helpers import get_test_config_path
+
+# pylint: enable=wrong-import-position
 
 # Config knobs shared by every family; small enough that construction is the only cost.
 _COMMON = {
@@ -82,6 +103,24 @@ _LAYERS = [
     ("gemma2", gemma2.Gemma2DecoderLayer, {"decoder_block": "gemma2"}, {}),
     ("qwen2", qwen2.Qwen2DecoderLayer, {"decoder_block": "qwen2"}, {"quant": None}),
     ("qwen3", qwen3.Qwen3DecoderLayer, {"decoder_block": "qwen3"}, {"quant": None}),
+]
+
+# Gemma 3 reads per-layer attention settings (the local/global pattern, rope scaling)
+# off the named model config, so it cannot run under the placeholder "default" name.
+# head_dim is pinned so it does not inherit the real 4b value and make the pass costly.
+_GEMMA3 = {
+    "model_name": "gemma3-4b",
+    "override_model_config": True,
+    "decoder_block": "gemma3",
+    "head_dim": 16,
+}
+
+# Layers onboarded to explicit sharding, as (case name, layer class, extra config).
+_SHARD_MODE_LAYERS = [
+    ("gemma", gemma.GemmaDecoderLayer, {"decoder_block": "gemma"}),
+    ("gemma2", gemma2.Gemma2DecoderLayer, {"decoder_block": "gemma2"}),
+    ("gemma3", gemma3.Gemma3DecoderLayer, _GEMMA3),
+    ("gemma3_scannable_block", gemma3.Gemma3ScannableBlock, _GEMMA3),
 ]
 
 
@@ -159,7 +198,11 @@ class DecoderLayerModelModeTest(parameterized.TestCase):
   def test_decode_call_overrides_prefill_construction(self, layer_cls, extra_config, ctor_kwargs):
     """Checks a layer built for prefill and called for decode, as MaxEngine does."""
     seen = self._mode_seen_by_attention(
-        layer_cls, extra_config, ctor_kwargs, MODEL_MODE_PREFILL, MODEL_MODE_AUTOREGRESSIVE
+        layer_cls,
+        extra_config,
+        ctor_kwargs,
+        MODEL_MODE_PREFILL,
+        MODEL_MODE_AUTOREGRESSIVE,
     )
     self.assertEqual(
         seen,
@@ -171,9 +214,99 @@ class DecoderLayerModelModeTest(parameterized.TestCase):
   def test_prefill_call_overrides_decode_construction(self, layer_cls, extra_config, ctor_kwargs):
     """Checks the reverse direction, so a layer cannot pass by hardcoding either mode."""
     seen = self._mode_seen_by_attention(
-        layer_cls, extra_config, ctor_kwargs, MODEL_MODE_AUTOREGRESSIVE, MODEL_MODE_PREFILL
+        layer_cls,
+        extra_config,
+        ctor_kwargs,
+        MODEL_MODE_AUTOREGRESSIVE,
+        MODEL_MODE_PREFILL,
     )
-    self.assertEqual(seen, MODEL_MODE_PREFILL, "attention ran in the layer's construction-time mode")
+    self.assertEqual(
+        seen,
+        MODEL_MODE_PREFILL,
+        "attention ran in the layer's construction-time mode",
+    )
+
+
+class DecoderLayerShardModeTest(parameterized.TestCase):
+  """Auto and explicit shard modes must agree on every onboarded decoder layer."""
+
+  def setUp(self):
+    super().setUp()
+    if jax.device_count() < 2:
+      self.skipTest(
+          "needs at least 2 devices for the mesh to shard anything; "
+          "set XLA_FLAGS=--xla_force_host_platform_device_count=8"
+      )
+
+  def _forward(self, layer_cls, extra_config, shard_mode):
+    """Builds one decoder layer in the given shard mode and runs a single forward pass.
+
+    The rng seeds are fixed so both modes initialize identical weights; any difference
+    in the returned activations is then attributable to the sharding annotations.
+
+    Args:
+      layer_cls: Decoder layer class to instantiate.
+      extra_config: Config overrides this family needs, on top of _COMMON.
+      shard_mode: Either "auto" or "explicit".
+
+    Returns:
+      The layer output as a numpy array.
+    """
+    cfg = pyconfig.initialize(
+        [sys.argv[0], get_test_config_path()],
+        **(_COMMON | extra_config | {"shard_mode": shard_mode}),
+    )
+    # Not _mesh(): explicit sharding needs the mesh axes typed Explicit rather than Auto,
+    # which only the config-driven constructor does.
+    mesh = maxtext_utils.get_mesh_from_config(cfg)
+
+    with nn_partitioning.axis_rules(cfg.logical_axis_rules), jax.set_mesh(mesh):
+      layer = layer_cls(
+          config=cfg,
+          mesh=mesh,
+          model_mode=MODEL_MODE_TRAIN,
+          rngs=nnx.Rngs(params=0, dropout=0),
+      )
+      batch = cfg.micro_batch_size_to_train_on
+      length = cfg.max_target_length
+      inputs = jnp.reshape(
+          jnp.arange(batch * length * cfg.emb_dim, dtype=jnp.float32) / (batch * length * cfg.emb_dim),
+          (batch, length, cfg.emb_dim),
+      )
+      positions = jnp.broadcast_to(jnp.arange(length, dtype=jnp.int32), (batch, length))
+      segment_ids = jnp.ones((batch, length), dtype=jnp.int32)
+      # Gemma3ScannableBlock returns the bare activations outside the scan path, while
+      # the single-layer classes always return an (output, kv_cache) pair.
+      output = layer(inputs, segment_ids, positions, True, MODEL_MODE_TRAIN)
+      if isinstance(output, tuple):
+        output = output[0]
+
+    return np.asarray(jax.device_get(output))
+
+  @parameterized.named_parameters(*_SHARD_MODE_LAYERS)
+  def test_explicit_matches_auto(self, layer_cls, extra_config):
+    """The explicit run must both succeed and reproduce the auto-mode activations."""
+    auto_out = self._forward(layer_cls, extra_config, "auto")
+    explicit_out = self._forward(layer_cls, extra_config, "explicit")
+
+    self.assertEqual(auto_out.shape, explicit_out.shape)
+    np.testing.assert_allclose(
+        explicit_out,
+        auto_out,
+        rtol=1e-5,
+        atol=1e-5,
+        err_msg="explicit shard mode changed the layer output; an out_sharding is wrong, not just missing",
+    )
+
+  @parameterized.named_parameters(*_SHARD_MODE_LAYERS)
+  def test_explicit_shard_mode_is_accepted_by_config(self, layer_cls, extra_config):
+    """Every onboarded decoder must be on the explicit-sharding allowlist in configs/types.py."""
+    del layer_cls
+    cfg = pyconfig.initialize(
+        [sys.argv[0], get_test_config_path()],
+        **(_COMMON | extra_config | {"shard_mode": "explicit"}),
+    )
+    self.assertEqual(cfg.shard_mode, ShardMode.EXPLICIT)
 
 
 if __name__ == "__main__":
